+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Le libellé de l'état `document-transmis-tdt` pour les actes est désormais "En attente du certificat RGS**" #1226
 - Passage en état `send-signature-error` s'il y a une erreur lors de l'envoi d'un dossier au parapheur FAST (flux studio et actes/helios) #967
 - Ajout d'un bouton dans le test du système permettant de vider le cache Redis #1257
+- Ajout des informations facultatives lors de l'utilisation du circuit à la volée pour signature avec le connecteur fast-parapheur #1259
 
 ## Corrections